worker/requirements.txt
numpy
pandas
psutil
pika
opencv-python
Pillow
fastparquet
pyarrow
setuptools
pydantic
python-dotenv
Deliver